Better organise legacy middleware
