+++ middleware/legacy/rules/static-pages.js
@@ -1,4 +1,10 @@
-export default {
+// Static page redirects
+
+import escapeRegExp from 'lodash/escapeRegExp';
+
+import { stringifyPathChunks } from '../utils';
+
+const redirects = {
   '/collections/world-war-I': '/entity/topic/83-1914-1918',
   '/collections/archaeology': '/entity/topic/80-archaeology',
   '/collections/art': '/entity/topic/190-art',
@@ -12,4 +18,21 @@ export default {
   '/collections/newspapers': '/entity/topic/18-newspaper',
   '/collections/photography': '/entity/topic/48-photography',
   '/collections/sport': '/entity/topic/114-sport'
+};
+
+
+export default (route) => {
+  for (const redirectFrom in redirects) {
+    const pattern = new RegExp(`^(/[a-z]{2})?${escapeRegExp(redirectFrom)}$`);
+    const match = route.path.match(pattern);
+    if (match) {
+      return {
+        path: stringifyPathChunks([
+          match[1],
+          redirects[redirectFrom]
+        ])
+      };
+    }
+  }
+  return null;
 };
