+++ middleware/legacy/utils.js
@@ -1,0 +1,3 @@
+export function stringifyPathChunks(chunks) {
+  return chunks.filter((chunk) => typeof chunk !== undefined).join('');
+}
